+++ src/packages/excalidraw/CHANGELOG.md
@@ -11,6 +11,18 @@ The change should be grouped under one of the below section and must contain PR 
 Please add the latest change on the top under the correct section.
 -->
 
+## Unreleased
+
+### Excalidraw API
+
+#### Refactor
+
+- Rename `appState.elementType` to [`appState.activeTool`](https://github.com/excalidraw/excalidraw/blob/master/src/types.ts#L80) which is now an object.
+
+##### BREAKING CHANGE
+
+You will need to pass `activeTool` instead of `elementType` from now onwards in `appState`
+
 ## 0.11.0 (2022-02-17)
 
 ## Excalidraw API